requirements.txt
openai
python-dotenv
langchain
tiktoken